Arreglados varios warnings

+++ src/LocoMotor/src/Transform.cpp
@@ -125,7 +125,7 @@ void LocoMotor::Transform::LookAt(const LMVector3& lookPos, const LMVector3& up)
 
 void LocoMotor::Transform::SetLocalPosition(const LMVector3& newPosition) {
 	_position = newPosition;
-	_gObjNode->SetPosition(newPosition.GetX(), newPosition.GetY(), newPosition.GetZ());
+	_gObjNode->SetPosition((float)newPosition.GetX(), (float) newPosition.GetY(), (float) newPosition.GetZ());
 }
 
 void LocoMotor::Transform::SetLocalRotation(const LMQuaternion& newRotation) {
@@ -137,7 +137,7 @@ void LocoMotor::Transform::SetLocalRotation(const LMQuaternion& newRotation) {
 
 void LocoMotor::Transform::SetLocalScale(const LMVector3& newSize) {
 	_scale = newSize;
-	_gObjNode->SetScale(newSize.GetX(), newSize.GetY(), newSize.GetZ());
+	_gObjNode->SetScale((float)newSize.GetX(), (float)newSize.GetY(), (float)newSize.GetZ());
 }
 
 void LocoMotor::Transform::SetPhysPosition(const LMVector3& newPosition) {
